LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/fullscreen/api/element-request-fullscreen-and-move-to-iframe-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Cannot request fullscreen because the associated document has changed.

PASS Element#requestFullscreen() followed by moving the element into an iframe
